[BBC-7096] ItemData.MDX
